Restore creation of capped collections
In order to start collections with specific options
(such as capped collections), it is necessary to
create collections explicitly.

The expression

    @client[collection, @collection_options].insert_many(records)

implicitly creates the collection without applying @collection_options.
This is the cause of issue #90.

This change checks to see if the collection to be written to
already exists in the database – if yes, it just remembers it.
If the collection does not exist, it is created according
to the configured plugin options. A new @collections
attribute helps to keep track of the collections already seen.